requirements.txt
coverage>=4.5
nose>=1.3
tornado>=2.21
numpy>=1.16
pandas==1.5.0
xlrd>=1.2
xlutils>=2.0
xlwt>=1.3
openpyxl==2.6
mock>=3.0
nose-exclude>=0.5
GitPython
loguru
